+++ setup.cfg
@@ -29,6 +29,7 @@ install_requires =
 where = src
 
 [flake8]
+docstring-convention=numpy
 ignore=
     # E800 Found commented out code
     E800,
@@ -36,6 +37,3 @@ ignore=
     E501,
     # W503 line break before binary operator
     W503
-
-[darglint]
-docstring_style=numpy
