+++ src/sw_solver/numpy.py
@@ -1,157 +1,12 @@
-"""Numpy version of the SWES solver."""  # noqa: DAR000
+"""Numpy version of the SWES solver."""
 
-import enum
-import math
-from dataclasses import dataclass
-from typing import Optional, Tuple, Type
+from typing import Optional, Tuple
 
 import numpy as np
 
-FloatT = Type[np.float64]
-
-FloatArray1D = Type[np.ndarray]
-FloatArray2D = Type[np.ndarray]
-FloatArray3D = Type[np.ndarray]
-
-
-@dataclass(frozen=True)
-class EarthConstants:
-    """
-    Constants for Earth.
-
-    g : gravity [m/s2]
-    rho: average atmosphere density [kg/m3]
-    a : average radius [m]
-    omega : rotation rate [Hz]
-    scaleHeight : atmosphere scale height [m]
-    nu : viscosity [m2/s]
-
-    """
-
-    g = 9.80616
-    rho = 1.2
-    a = 6.37122e6
-    omega = 7.292e-5
-    scaleHeight = 8.0e3
-    nu = 5.0e5
-
-
-EARTH_CONSTANTS = EarthConstants()
-
-
-class ICType(enum.IntEnum):
-    """Initial condition Types."""
-
-    RossbyHaurwitzWave = 0
-    """Test case 6 by Williamson."""
-
-    ZonalGeostrophicFlow = 1
-    """Test case 2 by Williamson."""
-
-
-class CartesianGrid:
-    """Cartesian Grid.
-
-    TODO: Document attributes.
-
-    """
-
-    x: FloatArray2D
-    y: FloatArray2D
-
-    dx: FloatArray2D
-    dy: FloatArray2D
-    dy1: FloatArray2D
-
-    dxmin: FloatT
-    dymin: FloatT
-
-    dxc: FloatArray2D
-    dyc: FloatArray2D
-    dy1c: FloatArray2D
-
-    def __init__(self, latlon_grid: "LatLonGrid"):
-        """Set up the grid."""
-        # Coordinates
-        self.x = EARTH_CONSTANTS.a * np.cos(latlon_grid.theta) * latlon_grid.phi
-        self.y = EARTH_CONSTANTS.a * latlon_grid.theta
-        y1 = EARTH_CONSTANTS.a * np.sin(latlon_grid.theta)
-
-        # Increments
-        self.dx = self.x[1:, :] - self.x[:-1, :]
-        self.dy = self.y[:, 1:] - self.y[:, :-1]
-        self.dy1 = y1[:, 1:] - y1[:, :-1]
-
-        # Compute mimimum distance between grid points on the sphere.
-        # This will be useful for CFL condition
-        self.dxmin = self.dx.min()
-        self.dymin = self.dy.min()
-
-        # "Centred" increments. Useful for updating solution
-        # with Lax-Wendroff scheme
-        self.dxc = 0.5 * (self.dx[:-1, 1:-1] + self.dx[1:, 1:-1])
-        self.dyc = 0.5 * (self.dy[1:-1, :-1] + self.dy[1:-1, 1:])
-        self.dy1c = 0.5 * (self.dy1[1:-1, :-1] + self.dy1[1:-1, 1:])
-
-
-class LatLonGrid:
-    """Latitude-Longitude Grid.
-
-    TODO: Document attributes.
-
-    """
-
-    phi: FloatArray2D
-    theta: FloatArray2D
-
-    c: FloatArray2D
-
-    cMidy: FloatArray2D
-
-    tg: FloatArray2D
-    tgMidx: FloatArray2D
-    tgMidy: FloatArray2D
-
-    def __init__(self, M: int, N: int):
-        """Set up the grid."""
-        has_points = (M > 1) and (N > 1)
-        if not has_points:
-            raise ValueError(
-                f"Number of grid points along each direction must be greater than one. Found {(M, N)}"
-            )
-
-        # Enforce an even number of points along latitude
-        N = N if N % 2 == 0 else N + 1
-
-        # Discretize longitude
-        dphi = 2.0 * math.pi / M
-        phi1D = np.linspace(-dphi, 2.0 * math.pi + dphi, M + 3)
-
-        # Discretize latitude
-        # Note: we exclude the poles and only consider a channel from -85 S to 85 N to avoid pole problem
-        # Note: the number of grid points must be even to prevent f to vanish
-        #       (important for computing initial height and velocity in geostrophic balance)
-        theta_range = 85.0
-        theta1D = np.linspace(
-            -theta_range / 180.0 * math.pi, theta_range / 180.0 * math.pi, N
-        )
-
-        # Build grid
-        self.phi, self.theta = np.meshgrid(phi1D, theta1D, indexing="ij")
-
-        # Cosine of mid-point values for theta along y
-        self.c = np.cos(self.theta)
-        self.cMidy = np.cos(0.5 * (self.theta[1:-1, 1:] + self.theta[1:-1, :-1]))
-
-        # Compute $\tan(\theta)$
-        self.tg = np.tan(self.theta[1:-1, 1:-1])
-        self.tgMidx = np.tan(0.5 * (self.theta[:-1, 1:-1] + self.theta[1:, 1:-1]))
-        self.tgMidy = np.tan(0.5 * (self.theta[1:-1, :-1] + self.theta[1:-1, 1:]))
-
-    @property
-    def shape(self) -> Tuple[int, int]:
-        """Return the grid size."""
-        return self.phi.shape
+from .grid import CartesianGrid, LatLonGrid
+from .ic import ICType, get_initial_conditions
+from .utils import EARTH_CONSTANTS, FloatArray1D, FloatArray2D, FloatArray3D, FloatT
 
 
 class DiffusionCoefficients:
@@ -206,146 +61,6 @@ class DiffusionCoefficients:
             cart_grid.dy[1:-1, :-1] * (cart_grid.dy[1:-1, 1:] + cart_grid.dy[1:-1, :-1])
         )
         self.Cy = np.concatenate((Cy[:, 0:1], Cy, Cy[:, -1:]), axis=1)
-
-
-def get_initial_conditions(ic_type: ICType, latlon_grid: LatLonGrid):
-    """
-    Compute the initial condition state based on the grid and type.
-
-    Parameters
-    ----------
-    ic_type : ICType
-        The initial condition type (see enum).
-    latlon_grid : LatLonGrid
-        Input grid on lat-lon points.
-
-    Returns
-    -------
-    h : FloatArray2D
-        Initial fluid height.
-    u : FloatArray2D
-        Initial longitudinal velocity.
-    v : FloatArray2D
-        Initial latitudinal velocity.
-    f : FloatArray2D
-        Coriolis parameter [Hz].
-
-    """
-    # Coriolis force
-    f = 2.0 * EARTH_CONSTANTS.omega * np.sin(latlon_grid.theta)
-
-    if ic_type == ICType.RossbyHaurwitzWave:
-        # --- IC 0: sixth test case taken from Williamson's suite --- #
-        # ---       Rossby-Haurwitz Wave                          --- #
-
-        # Set constants
-        w = 7.848e-6
-        K = 7.848e-6
-        h0 = 8e3
-        R = 4.0
-
-        # Compute initial fluid height
-        A = 0.5 * w * (2.0 * EARTH_CONSTANTS.omega + w) * (
-            np.cos(latlon_grid.theta) ** 2.0
-        ) + 0.25 * (K ** 2.0) * (np.cos(latlon_grid.theta) ** (2.0 * R)) * (
-            (R + 1.0) * (np.cos(latlon_grid.theta) ** 2.0)
-            + (2.0 * (R ** 2.0) - R - 2.0)
-            - 2.0 * (R ** 2.0) * (np.cos(latlon_grid.theta) ** (-2.0))
-        )
-        B = (
-            (2.0 * (EARTH_CONSTANTS.omega + w) * K)
-            / ((R + 1.0) * (R + 2.0))
-            * (np.cos(latlon_grid.theta) ** R)
-            * (
-                ((R ** 2.0) + 2.0 * R + 2.0)
-                - ((R + 1.0) ** 2.0) * (np.cos(latlon_grid.theta) ** 2.0)
-            )
-        )
-        C = (
-            0.25
-            * (K ** 2.0)
-            * (np.cos(latlon_grid.theta) ** (2.0 * R))
-            * ((R + 1.0) * (np.cos(latlon_grid.theta) ** 2.0) - (R + 2.0))
-        )
-
-        h = (
-            h0
-            + (
-                (EARTH_CONSTANTS.a ** 2.0) * A
-                + (EARTH_CONSTANTS.a ** 2.0) * B * np.cos(R * latlon_grid.phi)
-                + (EARTH_CONSTANTS.a ** 2.0) * C * np.cos(2.0 * R * latlon_grid.phi)
-            )
-            / EARTH_CONSTANTS.g
-        )
-
-        # Compute initial wind
-        u = EARTH_CONSTANTS.a * w * np.cos(
-            latlon_grid.theta
-        ) + EARTH_CONSTANTS.a * K * (np.cos(latlon_grid.theta) ** (R - 1.0)) * (
-            R * (np.sin(latlon_grid.theta) ** 2.0) - (np.cos(latlon_grid.theta) ** 2.0)
-        ) * np.cos(
-            R * latlon_grid.phi
-        )
-        v = (
-            -EARTH_CONSTANTS.a
-            * K
-            * R
-            * (np.cos(latlon_grid.theta) ** (R - 1.0))
-            * np.sin(latlon_grid.theta)
-            * np.sin(R * latlon_grid.phi)
-        )
-
-    elif ic_type == ICType.ZonalGeostrophicFlow:
-        # --- IC 1: second test case taken from Williamson's suite --- #
-        # ----      Steady State Nonlinear Zonal Geostrophic Flow  --- #
-
-        # Suggested values for $\alpha$ for second
-        # test cases of Williamson's suite:
-        # 	- 0
-        # 	- 0.05
-        # 	- pi/2 - 0.05
-        # 	- pi/2
-        alpha = math.pi / 2
-
-        # Set constants
-        u0 = 2.0 * math.pi * EARTH_CONSTANTS.a / (12.0 * 24.0 * 3600.0)
-        h0 = 2.94e4 / EARTH_CONSTANTS.g
-
-        # Make Coriolis parameter dependent on longitude and latitude
-        f = (
-            2.0
-            * EARTH_CONSTANTS.omega
-            * (
-                -np.cos(latlon_grid.phi) * np.cos(latlon_grid.theta) * np.sin(alpha)
-                + np.sin(latlon_grid.theta) * np.cos(alpha)
-            )
-        )
-
-        # Compute initial height
-        h = (
-            h0
-            - (EARTH_CONSTANTS.a * EARTH_CONSTANTS.omega * u0 + 0.5 * (u0 ** 2.0))
-            * (
-                (
-                    -np.cos(latlon_grid.phi) * np.cos(latlon_grid.theta) * np.sin(alpha)
-                    + np.sin(latlon_grid.theta) * np.cos(alpha)
-                )
-                ** 2.0
-            )
-            / EARTH_CONSTANTS.g
-        )
-
-        # Compute initial wind
-        u = u0 * (
-            np.cos(latlon_grid.theta) * np.cos(alpha)
-            + np.cos(latlon_grid.phi) * np.sin(latlon_grid.theta) * np.sin(alpha)
-        )
-        v = -u0 * np.sin(latlon_grid.phi) * np.sin(alpha)
-
-    else:
-        raise ValueError(f"Unknown ICType: {ic_type}")
-
-    return h, u, v, f
 
 
 class NumpySolver:
